+++ lib/cucumber/formatter/summary.rb
@@ -4,12 +4,13 @@ require 'cucumber/formatter/io'
 require 'cucumber/formatter/console'
 require 'cucumber/formatter/console_counts'
 require 'cucumber/formatter/console_issues'
+require 'cucumber/formatter/backtrace_filter'
 require 'cucumber/core/test/result'
 require 'cucumber/formatter/ast_lookup'
 
 module Cucumber
   module Formatter
-    # Summary formatter, outputting only feature / scenario titles
+    # Summary formatter, outputting feature / scenario titles plus failure details
     class Summary
       include Io
       include Console
@@ -20,6 +21,7 @@ module Cucumber
         @ast_lookup = AstLookup.new(config)
         @counts = ConsoleCounts.new(@config)
         @issues = ConsoleIssues.new(@config, @ast_lookup)
+        @failed_results = []
         @start_time = Time.now
 
         @config.on_event :test_case_started do |event|
@@ -31,9 +33,14 @@ module Cucumber
           print_result event.result
         end
 
+        @config.on_event :test_step_finished do |event|
+          collect_failed_result(event.test_step, event.result)
+        end
+
         @config.on_event :test_run_finished do |_event|
           duration = Time.now - @start_time
           @io.puts
+          print_elements(@failed_results, :failed, 'steps')
           print_statistics(duration, @config, @counts, @issues)
         end
       end
@@ -60,6 +67,13 @@ module Cucumber
 
       def print_result(result)
         @io.puts format_string(result, result.to_sym)
+      end
+
+      def collect_failed_result(test_step, result)
+        return if test_step.hook?
+
+        result = result.with_filtered_backtrace(Cucumber::Formatter::BacktraceFilter)
+        @failed_results << result if result.failed?
       end
     end
   end